db: find block number from txn number over Data API
