+++ silkworm/infra/common/binary_find_if.cpp
@@ -1,5 +1,5 @@
 /*
-   Copyright 2023 The Silkworm Authors
+   Copyright 2022 The Silkworm Authors
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -14,23 +14,23 @@
    limitations under the License.
 */
 
-#include "binary_search.hpp"
+#include "binary_find_if.hpp"
 
-namespace silkworm::rpc {
+namespace silkworm {
 
-Task<size_t> binary_search(size_t n, BinaryPredicate pred) {
+size_t binary_find_if(size_t n, absl::FunctionRef<bool(size_t)> f) {
     size_t i{0};
     size_t j{n};
     while (j > i) {
         const size_t count{j - i};
         const size_t m{i + count / 2};
-        if (co_await pred(m)) {
+        if (f(m)) {
             j = m;
         } else {
             i = m + 1;
         }
     }
-    co_return i;
+    return i;
 }
 
-}  // namespace silkworm::rpc
+}  // namespace silkworm
